+++ crates/threshold-signature-server/src/validator/api.rs
@@ -16,13 +16,277 @@
 use crate::{
     chain_api::{
         entropy::{self},
-        EntropyConfig,
+        get_api, get_rpc, EntropyConfig,
     },
-    helpers::{launch::FORBIDDEN_KEYS, substrate::query_chain},
+    get_signer_and_x25519_secret,
+    helpers::{
+        launch::{FORBIDDEN_KEYS, LATEST_BLOCK_NUMBER_RESHARE},
+        substrate::{get_stash_address, get_validators_info, query_chain, submit_transaction},
+    },
+    signing_client::{protocol_transport::open_protocol_connections, ProtocolErr},
     validator::errors::ValidatorErr,
+    AppState,
 };
-use std::str::FromStr;
-use subxt::{backend::legacy::LegacyRpcMethods, utils::AccountId32, OnlineClient};
+use axum::{body::Bytes, extract::State, http::StatusCode};
+use entropy_kvdb::kv_manager::{helpers::serialize as key_serialize, KvManager};
+pub use entropy_protocol::{
+    decode_verifying_key,
+    errors::ProtocolExecutionErr,
+    execute_protocol::{execute_protocol_generic, Channels, PairWrapper},
+    KeyParams, KeyShareWithAuxInfo, Listener, PartyId, SessionId, ValidatorInfo,
+};
+use entropy_shared::{OcwMessageReshare, NETWORK_PARENT_KEY, SETUP_TIMEOUT_SECONDS};
+use parity_scale_codec::{Decode, Encode};
+use rand_core::OsRng;
+use sp_core::Pair;
+use std::{collections::BTreeSet, str::FromStr, time::Duration};
+use subxt::{
+    backend::legacy::LegacyRpcMethods, ext::sp_core::sr25519, tx::PairSigner, utils::AccountId32,
+    OnlineClient,
+};
+use synedrion::{
+    make_key_resharing_session, sessions::SessionId as SynedrionSessionId, KeyResharingInputs,
+    NewHolder, OldHolder,
+};
+use tokio::time::timeout;
+
+/// HTTP POST endpoint called by the off-chain worker (propagation pallet) during network reshare.
+///
+/// The HTTP request takes a Parity SCALE encoded [OcwMessageReshare] which indicates which validator is joining
+///
+/// This will trigger the key reshare process.
+#[tracing::instrument(skip_all)]
+pub async fn new_reshare(
+    State(app_state): State<AppState>,
+    encoded_data: Bytes,
+) -> Result<StatusCode, ValidatorErr> {
+    let data = OcwMessageReshare::decode(&mut encoded_data.as_ref())?;
+    // TODO: validate message came from chain (check reshare block # against current block number) see #941
+    let api = get_api(&app_state.configuration.endpoint).await?;
+    let rpc = get_rpc(&app_state.configuration.endpoint).await?;
+    validate_new_reshare(&api, &rpc, &data, &app_state.kv_store).await?;
+    let signers_query = entropy::storage().staking_extension().signers();
+    let signers = query_chain(&api, &rpc, signers_query, None)
+        .await?
+        .ok_or_else(|| ValidatorErr::ChainFetch("Error getting signers"))?;
+
+    let next_signers_query = entropy::storage().staking_extension().signers();
+    let next_signers = query_chain(&api, &rpc, next_signers_query, None)
+        .await?
+        .ok_or_else(|| ValidatorErr::ChainFetch("Error getting next signers"))?;
+
+    let validators_info = get_validators_info(&api, &rpc, next_signers)
+        .await
+        .map_err(|e| ValidatorErr::UserError(e.to_string()))?;
+
+    let (signer, x25519_secret_key) = get_signer_and_x25519_secret(&app_state.kv_store)
+        .await
+        .map_err(|e| ValidatorErr::UserError(e.to_string()))?;
+
+    let verifying_key_query = entropy::storage().registry().jump_start_progress();
+    let verifying_key = query_chain(&api, &rpc, verifying_key_query, None)
+        .await?
+        .ok_or_else(|| ValidatorErr::ChainFetch("Parent verifying key error"))?
+        .verifying_key
+        .ok_or_else(|| ValidatorErr::OptionUnwrapError("Failed to get verifying key".to_string()))?
+        .0;
+
+    let decoded_verifying_key = decode_verifying_key(
+        &verifying_key
+            .clone()
+            .try_into()
+            .map_err(|_| ValidatorErr::Conversion("Verifying key conversion"))?,
+    )
+    .map_err(|e| ValidatorErr::VerifyingKeyError(e.to_string()))?;
+
+    let is_proper_signer = &validators_info
+        .iter()
+        .any(|validator_info| validator_info.tss_account == *signer.account_id());
+
+    if !is_proper_signer {
+        return Ok(StatusCode::MISDIRECTED_REQUEST);
+    }
+    // get old key if have it
+    let my_stash_address = get_stash_address(&api, &rpc, signer.account_id())
+        .await
+        .map_err(|e| ValidatorErr::UserError(e.to_string()))?;
+    let old_holder: Option<OldHolder<KeyParams, PartyId>> =
+        if data.new_signer == my_stash_address.encode() {
+            None
+        } else {
+            let kvdb_result = app_state.kv_store.kv().get(&hex::encode(NETWORK_PARENT_KEY)).await?;
+            let key_share: KeyShareWithAuxInfo =
+                entropy_kvdb::kv_manager::helpers::deserialize(&kvdb_result)
+                    .ok_or_else(|| ValidatorErr::KvDeserialize("Failed to load KeyShare".into()))?;
+            Some(OldHolder { key_share: key_share.0 })
+        };
+    let party_ids: BTreeSet<PartyId> =
+        validators_info.iter().cloned().map(|x| PartyId::new(x.tss_account)).collect();
+
+    let old_holders_info = get_validators_info(&api, &rpc, signers)
+        .await
+        .map_err(|e| ValidatorErr::UserError(e.to_string()))?;
+    let old_holders: BTreeSet<PartyId> =
+        old_holders_info.iter().cloned().map(|x| PartyId::new(x.tss_account)).collect();
+
+    let new_holder = NewHolder {
+        verifying_key: decoded_verifying_key,
+        // TODO: get from chain see #941
+        old_threshold: party_ids.len(),
+        old_holders,
+    };
+    let key_info_query = entropy::storage().parameters().signers_info();
+    let threshold = query_chain(&api, &rpc, key_info_query, None)
+        .await?
+        .ok_or_else(|| ValidatorErr::ChainFetch("Failed to get signers info"))?
+        .threshold;
+
+    let inputs = KeyResharingInputs {
+        old_holder,
+        new_holder: Some(new_holder),
+        new_holders: party_ids.clone(),
+        new_threshold: threshold as usize,
+    };
+
+    let session_id = SessionId::Reshare { verifying_key, block_number: data.block_number };
+    let account_id = AccountId32(signer.signer().public().0);
+    let session_id_hash = session_id.blake2(None)?;
+    let pair = PairWrapper(signer.signer().clone());
+
+    let mut converted_validator_info = vec![];
+    let mut tss_accounts = vec![];
+    for validator_info in validators_info {
+        let validator_info = ValidatorInfo {
+            x25519_public_key: validator_info.x25519_public_key,
+            ip_address: validator_info.ip_address,
+            tss_account: validator_info.tss_account.clone(),
+        };
+        converted_validator_info.push(validator_info.clone());
+        tss_accounts.push(validator_info.tss_account.clone());
+    }
+
+    let (rx_ready, rx_from_others, listener) =
+        Listener::new(converted_validator_info.clone(), &account_id);
+    app_state
+        .listener_state
+        .listeners
+        .lock()
+        .map_err(|_| ValidatorErr::SessionError("Error getting lock".to_string()))?
+        .insert(session_id.clone(), listener);
+
+    open_protocol_connections(
+        &converted_validator_info,
+        &session_id,
+        signer.signer(),
+        &app_state.listener_state,
+        &x25519_secret_key,
+    )
+    .await?;
+
+    let channels = {
+        let ready = timeout(Duration::from_secs(SETUP_TIMEOUT_SECONDS), rx_ready).await?;
+        let broadcast_out = ready??;
+        Channels(broadcast_out, rx_from_others)
+    };
+
+    let session = make_key_resharing_session(
+        &mut OsRng,
+        SynedrionSessionId::from_seed(session_id_hash.as_slice()),
+        pair,
+        &party_ids,
+        inputs,
+    )
+    .map_err(ProtocolExecutionErr::SessionCreation)?;
+
+    let new_key_share = execute_protocol_generic(channels, session, session_id_hash)
+        .await
+        .map_err(|_| ValidatorErr::ProtocolError("Error executing protocol".to_string()))?
+        .0
+        .ok_or(ValidatorErr::NoOutputFromReshareProtocol)?;
+    let serialized_key_share = key_serialize(&new_key_share)
+        .map_err(|_| ProtocolErr::KvSerialize("Kv Serialize Error".to_string()))?;
+    let network_parent_key = hex::encode(NETWORK_PARENT_KEY);
+    // TODO: should this be a two step process? see # https://github.com/entropyxyz/entropy-core/issues/968
+    if app_state.kv_store.kv().exists(&network_parent_key).await? {
+        app_state.kv_store.kv().delete(&network_parent_key).await?
+    };
+
+    let reservation = app_state.kv_store.kv().reserve_key(network_parent_key).await?;
+    app_state.kv_store.kv().put(reservation, serialized_key_share.clone()).await?;
+
+    // TODO: Error handling really complex needs to be thought about.
+    confirm_key_reshare(&api, &rpc, &signer).await?;
+    Ok(StatusCode::OK)
+}
+
+// Validates new reshare endpoint
+/// Checks the chain for validity of data and block number of data matches current block
+pub async fn validate_new_reshare(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    chain_data: &OcwMessageReshare,
+    kv_manager: &KvManager,
+) -> Result<(), ValidatorErr> {
+    let last_block_number_recorded = kv_manager.kv().get(LATEST_BLOCK_NUMBER_RESHARE).await?;
+    if u32::from_be_bytes(
+        last_block_number_recorded
+            .try_into()
+            .map_err(|_| ValidatorErr::Conversion("Block number conversion"))?,
+    ) >= chain_data.block_number
+    {
+        return Err(ValidatorErr::RepeatedData);
+    }
+
+    let latest_block_number = rpc
+        .chain_get_header(None)
+        .await?
+        .ok_or_else(|| ValidatorErr::OptionUnwrapError("Failed to get block number".to_string()))?
+        .number;
+
+    // we subtract 1 as the message info is coming from the previous block
+    if latest_block_number.saturating_sub(1) != chain_data.block_number {
+        return Err(ValidatorErr::StaleData);
+    }
+
+    let reshare_data_info_query = entropy::storage().staking_extension().reshare_data();
+    let reshare_data = query_chain(api, rpc, reshare_data_info_query, None)
+        .await?
+        .ok_or_else(|| ValidatorErr::ChainFetch("Not Currently in a reshare"))?;
+
+    if reshare_data.new_signer != chain_data.new_signer
+        || chain_data.block_number != reshare_data.block_number
+    {
+        return Err(ValidatorErr::InvalidData);
+    }
+    kv_manager.kv().delete(LATEST_BLOCK_NUMBER_RESHARE).await?;
+    let reservation = kv_manager.kv().reserve_key(LATEST_BLOCK_NUMBER_RESHARE.to_string()).await?;
+    kv_manager.kv().put(reservation, chain_data.block_number.to_be_bytes().to_vec()).await?;
+
+    Ok(())
+}
+
+/// Confirms that a validator has succefully reshared.
+pub async fn confirm_key_reshare(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    signer: &PairSigner<EntropyConfig, sr25519::Pair>,
+) -> Result<(), ValidatorErr> {
+    // TODO error handling + return error
+    // TODO fire and forget, or wait for in block maybe Ddos error
+    // TODO: Understand this better, potentially use sign_and_submit_default
+    // or other method under sign_and_*
+    let block_hash = rpc
+        .chain_get_block_hash(None)
+        .await?
+        .ok_or_else(|| ValidatorErr::OptionUnwrapError("Error getting block hash".to_string()))?;
+
+    let nonce_call = entropy::apis().account_nonce_api().account_nonce(signer.account_id().clone());
+    let nonce = api.runtime_api().at(block_hash).call(nonce_call).await?;
+
+    let confirm_key_reshare_request = entropy::tx().staking_extension().confirm_key_reshare();
+    submit_transaction(api, rpc, signer, &confirm_key_reshare_request, Some(nonce)).await?;
+    Ok(())
+}
 
 /// Validation for if an account can cover tx fees for a tx
 pub async fn check_balance_for_fees(