+++ crates/threshold-signature-server/src/validator/errors.rs
@@ -15,12 +15,15 @@
 
 use std::string::FromUtf8Error;
 
+use crate::signing_client::ProtocolErr;
 use axum::{
     http::StatusCode,
     response::{IntoResponse, Response},
 };
-use entropy_protocol::sign_and_encrypt::EncryptedSignedMessageErr;
+use entropy_protocol::{errors::ProtocolExecutionErr, sign_and_encrypt::EncryptedSignedMessageErr};
+use synedrion::sessions;
 use thiserror::Error;
+use tokio::sync::oneshot::error::RecvError;
 
 #[derive(Debug, Error)]
 pub enum ValidatorErr {
@@ -52,6 +55,44 @@ pub enum ValidatorErr {
     Authentication,
     #[error("Substrate: {0}")]
     SubstrateClient(#[from] entropy_client::substrate::SubstrateError),
+    #[error("Codec decoding error: {0}")]
+    CodecError(#[from] parity_scale_codec::Error),
+    #[error("User Error: {0}")]
+    UserError(String),
+    #[error("Option Unwrap error: {0}")]
+    OptionUnwrapError(String),
+    #[error("Vec<u8> Conversion Error: {0}")]
+    Conversion(&'static str),
+    #[error("Verifying key Error: {0}")]
+    VerifyingKeyError(String),
+    #[error("Session Error: {0}")]
+    SessionError(String),
+    #[error("Protocol Execution Error {0}")]
+    ProtocolExecution(#[from] ProtocolExecutionErr),
+    #[error("Listener: {0}")]
+    Listener(#[from] entropy_protocol::errors::ListenerErr),
+    #[error("Reshare protocol error: {0}")]
+    SigningClientError(#[from] ProtocolErr),
+    #[error("Timed out waiting for remote party")]
+    Timeout(#[from] tokio::time::error::Elapsed),
+    #[error("Oneshot timeout error: {0}")]
+    OneshotTimeout(#[from] RecvError),
+    #[error("Synedrion session creation error: {0}")]
+    SessionCreation(sessions::LocalError),
+    #[error("No output from reshare protocol")]
+    NoOutputFromReshareProtocol,
+    #[error("Protocol Error: {0}")]
+    ProtocolError(String),
+    #[error("Kv Fatal error")]
+    KvSerialize(String),
+    #[error("Kv Deserialization Error: {0}")]
+    KvDeserialize(String),
+    #[error("Data is stale")]
+    StaleData,
+    #[error("Data is not verifiable")]
+    InvalidData,
+    #[error("Data is repeated")]
+    RepeatedData,
 }
 
 impl IntoResponse for ValidatorErr {