+++ crates/threshold-signature-server/src/validator/tests.rs
@@ -12,19 +12,179 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-use entropy_shared::MIN_BALANCE;
-use entropy_testing_utils::{
-    constants::{ALICE_STASH_ADDRESS, RANDOM_ACCOUNT},
-    substrate_context::testing_context,
-};
-
 use super::api::{check_balance_for_fees, check_forbidden_key};
 use crate::{
-    chain_api::{get_api, get_rpc},
-    helpers::{launch::FORBIDDEN_KEYS, tests::initialize_test_logger},
-    validator::errors::ValidatorErr,
+    chain_api::{
+        entropy::{self, runtime_types::bounded_collections::bounded_vec},
+        get_api, get_rpc, EntropyConfig,
+    },
+    helpers::{
+        launch::{
+            development_mnemonic, ValidatorName, FORBIDDEN_KEYS, LATEST_BLOCK_NUMBER_RESHARE,
+        },
+        substrate::submit_transaction,
+        tests::{
+            initialize_test_logger, run_to_block, setup_client, spawn_testing_validators,
+            unsafe_get,
+        },
+        validator::get_signer_and_x25519_secret_from_mnemonic,
+    },
+    validator::{api::validate_new_reshare, errors::ValidatorErr},
+};
+use entropy_kvdb::clean_tests;
+use entropy_shared::{
+    OcwMessageReshare, EVE_VERIFYING_KEY, MIN_BALANCE, NETWORK_PARENT_KEY,
+    TEST_RESHARE_BLOCK_NUMBER,
+};
+use entropy_testing_utils::{
+    constants::{ALICE_STASH_ADDRESS, RANDOM_ACCOUNT},
+    substrate_context::{test_node_process_testing_state, testing_context},
+    test_context_stationary,
+};
+use futures::future::join_all;
+use parity_scale_codec::Encode;
+use serial_test::serial;
+use sp_keyring::AccountKeyring;
+use subxt::{
+    backend::legacy::LegacyRpcMethods, ext::sp_core::sr25519, tx::PairSigner, OnlineClient,
 };
 
+#[tokio::test]
+#[serial]
+async fn test_reshare() {
+    initialize_test_logger().await;
+    clean_tests();
+
+    let alice = AccountKeyring::Alice;
+
+    let cxt = test_node_process_testing_state(true).await;
+    let (_validator_ips, _validator_ids) = spawn_testing_validators(true).await;
+    let validator_ports = vec![3001, 3002, 3003];
+    let api = get_api(&cxt.ws_url).await.unwrap();
+    let rpc = get_rpc(&cxt.ws_url).await.unwrap();
+
+    let client = reqwest::Client::new();
+    let mut key_shares_before = vec![];
+    for port in &validator_ports {
+        key_shares_before.push(unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), *port).await);
+    }
+
+    setup_for_reshare(&api, &rpc).await;
+
+    let block_number = TEST_RESHARE_BLOCK_NUMBER;
+    let onchain_reshare_request =
+        OcwMessageReshare { new_signer: alice.public().encode(), block_number };
+
+    run_to_block(&rpc, block_number + 1).await;
+
+    let response_results = join_all(
+        validator_ports
+            .iter()
+            .map(|port| {
+                client
+                    .post(format!("http://127.0.0.1:{}/validator/reshare", port))
+                    .body(onchain_reshare_request.clone().encode())
+                    .send()
+            })
+            .collect::<Vec<_>>(),
+    )
+    .await;
+    for response_result in response_results {
+        assert_eq!(response_result.unwrap().text().await.unwrap(), "");
+    }
+    for i in 0..validator_ports.len() {
+        assert_ne!(
+            key_shares_before[i],
+            unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), validator_ports[i]).await
+        );
+    }
+    clean_tests();
+}
+
+#[tokio::test]
+#[serial]
+async fn test_reshare_validation_fail() {
+    initialize_test_logger().await;
+    clean_tests();
+
+    let dave = AccountKeyring::Dave;
+    let cxt = test_node_process_testing_state(true).await;
+    let api = get_api(&cxt.ws_url).await.unwrap();
+    let rpc = get_rpc(&cxt.ws_url).await.unwrap();
+    let kv = setup_client().await;
+
+    let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
+    let mut ocw_message = OcwMessageReshare { new_signer: dave.public().encode(), block_number };
+
+    let err_stale_data =
+        validate_new_reshare(&api, &rpc, &ocw_message, &kv).await.map_err(|e| e.to_string());
+    assert_eq!(err_stale_data, Err("Data is stale".to_string()));
+
+    let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
+    ocw_message.block_number = block_number;
+    run_to_block(&rpc, block_number + 1).await;
+
+    let err_incorrect_data =
+        validate_new_reshare(&api, &rpc, &ocw_message, &kv).await.map_err(|e| e.to_string());
+    assert_eq!(err_incorrect_data, Err("Data is not verifiable".to_string()));
+
+    // manipulates kvdb to get to repeated data error
+    kv.kv().delete(LATEST_BLOCK_NUMBER_RESHARE).await.unwrap();
+    let reservation = kv.kv().reserve_key(LATEST_BLOCK_NUMBER_RESHARE.to_string()).await.unwrap();
+    kv.kv().put(reservation, (block_number + 5).to_be_bytes().to_vec()).await.unwrap();
+
+    let err_stale_data =
+        validate_new_reshare(&api, &rpc, &ocw_message, &kv).await.map_err(|e| e.to_string());
+    assert_eq!(err_stale_data, Err("Data is repeated".to_string()));
+    clean_tests();
+}
+
+#[tokio::test]
+#[serial]
+async fn test_reshare_validation_fail_not_in_reshare() {
+    initialize_test_logger().await;
+    clean_tests();
+
+    let alice = AccountKeyring::Alice;
+    let cxt = test_context_stationary().await;
+    let api = get_api(&cxt.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&cxt.node_proc.ws_url).await.unwrap();
+    let kv = setup_client().await;
+
+    let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
+    let ocw_message = OcwMessageReshare { new_signer: alice.public().encode(), block_number };
+
+    run_to_block(&rpc, block_number + 1).await;
+
+    let err_not_in_reshare =
+        validate_new_reshare(&api, &rpc, &ocw_message, &kv).await.map_err(|e| e.to_string());
+    assert_eq!(err_not_in_reshare, Err("Chain Fetch: Not Currently in a reshare".to_string()));
+
+    clean_tests();
+}
+
+async fn setup_for_reshare(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+) {
+    let alice = AccountKeyring::Alice;
+    let signer = PairSigner::<EntropyConfig, sr25519::Pair>::new(alice.clone().into());
+
+    let jump_start_request = entropy::tx().registry().jump_start_network();
+    let _result = submit_transaction(api, rpc, &signer, &jump_start_request, None).await.unwrap();
+
+    let validators_names = vec![ValidatorName::Alice, ValidatorName::Bob, ValidatorName::Charlie];
+    for validator_name in validators_names {
+        let mnemonic = development_mnemonic(&Some(validator_name));
+        let (tss_signer, _static_secret) =
+            get_signer_and_x25519_secret_from_mnemonic(&mnemonic.to_string()).unwrap();
+        let jump_start_confirm_request = entropy::tx()
+            .registry()
+            .confirm_jump_start(bounded_vec::BoundedVec(EVE_VERIFYING_KEY.to_vec()));
+
+        submit_transaction(api, rpc, &tss_signer, &jump_start_confirm_request, None).await.unwrap();
+    }
+}
 #[tokio::test]
 #[should_panic = "Account does not exist, add balance"]
 async fn test_check_balance_for_fees() {