+++ pallets/registry/src/benchmarking.rs
@@ -14,7 +14,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 //! Benchmarking setup for pallet-propgation
-use entropy_shared::{SIGNING_PARTY_SIZE, VERIFICATION_KEY_LENGTH};
+use entropy_shared::{TOTAL_SIGNERS, VERIFICATION_KEY_LENGTH};
 use frame_benchmarking::{account, benchmarks, impl_benchmark_test_suite, whitelisted_caller};
 use frame_support::{
     traits::{Currency, Get},
@@ -79,26 +79,26 @@ benchmarks! {
   }
 
   confirm_jump_start_done {
-    let c in 0 .. SIGNING_PARTY_SIZE as u32;
+    let c in 0 .. TOTAL_SIGNERS as u32;
     let sig_req_account: T::AccountId = whitelisted_caller();
     let validator_account: T::AccountId = whitelisted_caller();
     let expected_verifying_key = BoundedVec::default();
 
     let mut accounts = vec![];
-    for i in 0..SIGNING_PARTY_SIZE {
+    for i in 0..TOTAL_SIGNERS {
         accounts.push(account::<T::AccountId>("ts_account", i as u32, SEED));
     }
 
-    let validators = add_non_syncing_validators::<T>(SIGNING_PARTY_SIZE as u32, 0);
+    let validators = add_non_syncing_validators::<T>(TOTAL_SIGNERS as u32, 0);
     <Validators<T>>::set(validators.clone());
 
-    for i in 0..SIGNING_PARTY_SIZE {
-        <ThresholdToStash<T>>::insert(accounts[i].clone(), &validators[i]);
+    for i in 0..TOTAL_SIGNERS {
+        <ThresholdToStash<T>>::insert(accounts[i as usize].clone(), &validators[i as usize]);
     }
 
     <JumpStartProgress<T>>::put(JumpStartDetails {
       jump_start_status: JumpStartStatus::InProgress(0),
-      confirmations: vec![validators[0].clone()],
+      confirmations: vec![validators[0].clone(), validators[0].clone()],
       verifying_key: None
       });
 
@@ -111,17 +111,17 @@ benchmarks! {
   }
 
   confirm_jump_start_confirm {
-    let c in 0 .. SIGNING_PARTY_SIZE as u32;
+    let c in 0 .. TOTAL_SIGNERS as u32;
     let sig_req_account: T::AccountId = whitelisted_caller();
     let validator_account: T::AccountId = whitelisted_caller();
     let threshold_account: T::AccountId = whitelisted_caller();
     let expected_verifying_key = BoundedVec::default();
 
     // add validators and a registering user
-    for i in 0..SIGNING_PARTY_SIZE {
-        let validators = add_non_syncing_validators::<T>(SIGNING_PARTY_SIZE as u32, 0);
+    for i in 0..TOTAL_SIGNERS {
+        let validators = add_non_syncing_validators::<T>(TOTAL_SIGNERS as u32, 0);
         <Validators<T>>::set(validators.clone());
-        <ThresholdToStash<T>>::insert(&threshold_account, &validators[i]);
+        <ThresholdToStash<T>>::insert(&threshold_account, &validators[i as usize]);
     }
     <JumpStartProgress<T>>::put(JumpStartDetails {
       jump_start_status: JumpStartStatus::InProgress(0),