+++ pallets/staking/src/weights.rs
@@ -57,7 +57,8 @@ pub trait WeightInfo {
 	fn withdraw_unbonded() -> Weight;
 	fn validate() -> Weight;
 	fn declare_synced() -> Weight;
-	fn new_session_handler_helper(c: u32, n: u32, ) -> Weight;
+	fn confirm_key_reshare_confirmed(c: u32) -> Weight;
+	fn confirm_key_reshare_completed() -> Weight;
 }
 
 /// Weights for pallet_staking_extension using the Substrate node and recommended hardware.
@@ -158,23 +159,36 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 			.saturating_add(T::DbWeight::get().reads(1_u64))
 			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
-	/// Storage: `StakingExtension::SigningGroups` (r:2 w:2)
-	/// Proof: `StakingExtension::SigningGroups` (`max_values`: None, `max_size`: None, mode: `Measured`)
-	/// The range of component `c` is `[0, 1000]`.
-	/// The range of component `n` is `[0, 1000]`.
-	fn new_session_handler_helper(c: u32, n: u32, ) -> Weight {
+	/// Storage: `StakingExtension::ThresholdToStash` (r:1 w:0)
+	/// Proof: `StakingExtension::ThresholdToStash` (`max_values`: None, `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::NextSigners` (r:1 w:1)
+	/// Proof: `StakingExtension::NextSigners` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	/// The range of component `c` is `[0, 2]`.
+	fn confirm_key_reshare_confirmed(_c: u32, ) -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `218 + c * (32 ±0)`
-		//  Estimated: `6156 + c * (32 ±0)`
-		// Minimum execution time: 20_000_000 picoseconds.
-		Weight::from_parts(20_000_000, 6156)
-			// Standard Error: 24_941
-			.saturating_add(Weight::from_parts(798_680, 0).saturating_mul(c.into()))
-			// Standard Error: 24_941
-			.saturating_add(Weight::from_parts(821_922, 0).saturating_mul(n.into()))
-			.saturating_add(T::DbWeight::get().reads(2_u64))
-			.saturating_add(T::DbWeight::get().writes(2_u64))
-			.saturating_add(Weight::from_parts(0, 32).saturating_mul(c.into()))
+		//  Measured:  `445`
+		//  Estimated: `3910`
+		// Minimum execution time: 11_000_000 picoseconds.
+		Weight::from_parts(11_916_666, 0)
+			.saturating_add(Weight::from_parts(0, 3910))
+			.saturating_add(T::DbWeight::get().reads(2))
+			.saturating_add(T::DbWeight::get().writes(1))
+	}
+	/// Storage: `StakingExtension::ThresholdToStash` (r:1 w:0)
+	/// Proof: `StakingExtension::ThresholdToStash` (`max_values`: None, `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::NextSigners` (r:1 w:1)
+	/// Proof: `StakingExtension::NextSigners` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::Signers` (r:0 w:1)
+	/// Proof: `StakingExtension::Signers` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	fn confirm_key_reshare_completed() -> Weight {
+		// Proof Size summary in bytes:
+		//  Measured:  `477`
+		//  Estimated: `3942`
+		// Minimum execution time: 12_000_000 picoseconds.
+		Weight::from_parts(13_000_000, 0)
+			.saturating_add(Weight::from_parts(0, 3942))
+			.saturating_add(T::DbWeight::get().reads(2))
+			.saturating_add(T::DbWeight::get().writes(2))
 	}
 }
 
@@ -275,22 +289,35 @@ impl WeightInfo for () {
 			.saturating_add(RocksDbWeight::get().reads(1_u64))
 			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
-	/// Storage: `StakingExtension::SigningGroups` (r:2 w:2)
-	/// Proof: `StakingExtension::SigningGroups` (`max_values`: None, `max_size`: None, mode: `Measured`)
-	/// The range of component `c` is `[0, 1000]`.
-	/// The range of component `n` is `[0, 1000]`.
-	fn new_session_handler_helper(c: u32, n: u32, ) -> Weight {
+	/// Storage: `StakingExtension::ThresholdToStash` (r:1 w:0)
+	/// Proof: `StakingExtension::ThresholdToStash` (`max_values`: None, `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::NextSigners` (r:1 w:1)
+	/// Proof: `StakingExtension::NextSigners` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	/// The range of component `c` is `[0, 2]`.
+	fn confirm_key_reshare_confirmed(_c: u32, ) -> Weight {
 		// Proof Size summary in bytes:
-		//  Measured:  `218 + c * (32 ±0)`
-		//  Estimated: `6156 + c * (32 ±0)`
-		// Minimum execution time: 20_000_000 picoseconds.
-		Weight::from_parts(20_000_000, 6156)
-			// Standard Error: 24_941
-			.saturating_add(Weight::from_parts(798_680, 0).saturating_mul(c.into()))
-			// Standard Error: 24_941
-			.saturating_add(Weight::from_parts(821_922, 0).saturating_mul(n.into()))
-			.saturating_add(RocksDbWeight::get().reads(2_u64))
-			.saturating_add(RocksDbWeight::get().writes(2_u64))
-			.saturating_add(Weight::from_parts(0, 32).saturating_mul(c.into()))
+		//  Measured:  `445`
+		//  Estimated: `3910`
+		// Minimum execution time: 11_000_000 picoseconds.
+		Weight::from_parts(11_916_666, 0)
+			.saturating_add(Weight::from_parts(0, 3910))
+			.saturating_add(RocksDbWeight::get().reads(2))
+			.saturating_add(RocksDbWeight::get().writes(1))
+	}
+	/// Storage: `StakingExtension::ThresholdToStash` (r:1 w:0)
+	/// Proof: `StakingExtension::ThresholdToStash` (`max_values`: None, `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::NextSigners` (r:1 w:1)
+	/// Proof: `StakingExtension::NextSigners` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	/// Storage: `StakingExtension::Signers` (r:0 w:1)
+	/// Proof: `StakingExtension::Signers` (`max_values`: Some(1), `max_size`: None, mode: `Measured`)
+	fn confirm_key_reshare_completed() -> Weight {
+		// Proof Size summary in bytes:
+		//  Measured:  `477`
+		//  Estimated: `3942`
+		// Minimum execution time: 12_000_000 picoseconds.
+		Weight::from_parts(13_000_000, 0)
+			.saturating_add(Weight::from_parts(0, 3942))
+			.saturating_add(RocksDbWeight::get().reads(2))
+			.saturating_add(RocksDbWeight::get().writes(2))
 	}
 }